# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Smart Monitor API",
    "description": "Complete API documentation for Smart Monitor - Distributed System Monitoring Platform\n\n## Features\n- Agent Registration & Authentication\n- Real-time Metrics Streaming\n- System Monitoring (CPU, RAM, Disk)\n- Health & Readiness Checks\n\n## Quick Start\n1. Register agent: POST /v1/agent/register\n2. Stream metrics: POST /v1/stats/stream\n3. Query stats: GET /v1/stats/{hostname}\n\n## Authentication\nMost endpoints require access token obtained during registration.",
    "version": "1.0.0",
    "contact": {
      "name": "Smart Monitor Team",
      "email": "support@smartmonitor.example.com",
      "url": "https://smartmonitor.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "access_token",
      "in": "query",
      "description": "Access token obtained from agent registration"
    },
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer token format: 'Bearer {access_token}'"
    }
  },
...
  "paths": {
...
    "/v1/stats/stream": {
      "post": {
        "tags": ["Metrics"],
        "summary": "Stream system metrics",
        "description": "Stream real-time system metrics. Requires valid access token. Recommended interval: 2-10 seconds.",
        "operationId": "streamStats",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [
          {"ApiKeyAuth": []}
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "System metrics data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StatsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metrics recorded successfully",
            "schema": {
              "$ref": "#/definitions/StatsResponse"
            },
            "examples": {
              "application/json": {
                "message": "Stats recorded successfully",
                "timestamp": 1737882600
              }
            }
          },
          "400": {
            "description": "Invalid metrics data",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            },
            "examples": {
              "application/json": {
                "message": "Invalid stats data: CPU value must be between 0 and 100",
                "code": "VALIDATION_ERROR"
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            },
            "examples": {
              "application/json": {
                "message": "Authentication failed: invalid or expired token",
                "code": "UNAUTHORIZED"
              }
            }
          }
        }
      }
    },
    "/v1/stats/{hostname}": {
      "get": {
        "tags": ["Metrics"],
        "summary": "Get stats for specific host",
        "description": "Retrieve latest metrics for a hostname",
        "operationId": "getStats",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "hostname",
            "in": "path",
            "description": "Hostname to query",
            "required": true,
            "type": "string",
            "default": "server-01"
          }
        ],
        "responses": {
          "200": {
            "description": "Stats retrieved successfully",
            "schema": {
              "$ref": "#/definitions/StatsDetailResponse"
            },
            "examples": {
              "application/json": {
                "hostname": "server-01",
                "agent_id": "agent-a3f5c2d1",
                "ip_address": "192.168.1.10",
                "cpu": 45.2,
                "ram": 68.5,
                "disk": 72.3,
                "timestamp": 1737882600,
                "last_received": "2026-01-26T10:30:00Z",
                "status": "online"
              }
            }
          },
          "404": {
            "description": "Host not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "StatsRequest": {
      "type": "object",
      "required": ["hostname", "agent_id", "access_token", "cpu", "ram", "disk"],
      "properties": {
        "hostname": {
          "type": "string",
          "example": "server-01"
        },
        "agent_id": {
          "type": "string",
          "example": "agent-a3f5c2d1"
        },
        "access_token": {
          "type": "string",
          "example": "3f4a8b2c1d9e7f6a5b4c3d2e1f0a9b8c7d6e5f4a3b2c1d0e9f8a7b6c5d4e3f2"
        },
        "ip_address": {
          "type": "string",
          "example": "192.168.1.10"
        },
        "agent_version": {
          "type": "string",
          "example": "1.0.0"
        },
        "cpu": {
          "type": "number",
          "format": "double",
          "minimum": 0,
          "maximum": 100,
          "description": "CPU usage percentage",
          "example": 45.2
        },
        "ram": {
          "type": "number",
          "format": "double",
          "minimum": 0,
          "maximum": 100,
          "description": "RAM usage percentage",
          "example": 68.5
        },
        "disk": {
          "type": "number",
          "format": "double",
          "minimum": 0,
          "maximum": 100,
          "description": "Disk usage percentage",
          "example": 72.3
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "StatsResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Stats recorded successfully"
        },
        "timestamp": {
          "type": "integer",
          "format": "int64",
          "example": 1737882600
        }
      }
    },
    "StatsDetailResponse": {
      "type": "object",
      "properties": {
        "hostname": {
          "type": "string",
          "example": "server-01"
        },
        "agent_id": {
          "type": "string",
          "example": "agent-a3f5c2d1"
        },
        "ip_address": {
          "type": "string",
          "example": "192.168.1.10"
        },
        "cpu": {
          "type": "number",
          "format": "double",
          "example": 45.2
        },
        "ram": {
          "type": "number",
          "format": "double",
          "example": 68.5
        },
        "disk": {
          "type": "number",
          "format": "double",
          "example": 72.3
        },
        "timestamp": {
          "type": "integer",
          "format": "int64",
          "example": 1737882600
        },
        "last_received": {
          "type": "string",
          "format": "date-time",
          "example": "2026-01-26T10:30:00Z"
        },
        "status": {
          "type": "string",
          "enum": ["online", "offline", "unknown"],
          "example": "online"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "description": "Error description",
          "example": "Invalid request parameters"
        },
        "code": {
          "type": "string",
          "description": "Error code",
          "enum": [
            "INVALID_REQUEST",
            "UNAUTHORIZED",
            "NOT_FOUND",
            "VALIDATION_ERROR",
            "INTERNAL_ERROR"
          ],
          "example": "INVALID_REQUEST"
        },
        "details": {
          "type": "object",
          "description": "Additional error details"
        }
      }
    }
  },
...
}